+++ package.json
@@ -22,6 +22,7 @@
     "migrations:down": "cd ./packages/backend/migrations && bun run migrations:down",
     "migrations:generate": "cd ./packages/backend/migrations && bun run migrations:generate",
     "compose:up": "docker compose -f ./deployment/docker-compose.yml up -d",
+    "compose:build": "docker compose -f ./deployment/docker-compose.yml up --build",
     "compose:down": "docker compose -f ./deployment/docker-compose.yml down",
     "dev:backend:api": "cd ./packages/backend/api && bun run dev",
     "dev:frontend:web": "cd ./packages/frontend/web && bun run dev",
